Fixed VR postfx rendering.
